requirements.txt
torch>=1.10.0
numpy>=1.20.0
cn_clip
ultralytics>=8.0.0
pillow>=8.0.0
tqdm
scikit-learn
matplotlib
librosa>=0.9.0
transformers>=4.18.0
soundfile
tensorboard
pandas